+++ sound/src/main/java/com/kylecorry/andromeda/sound/ToneGenerator.kt
@@ -7,7 +7,7 @@ class ToneGenerator {
 
     private val soundGenerator = SoundGenerator()
 
-    fun getTone(frequency: Int, sampleRate: Int = 64000, durationSeconds: Int = 1): AudioTrack {
+    fun getTone(frequency: Int, sampleRate: Int = 64000, durationSeconds: Float = 1f): AudioTrack {
         return soundGenerator.getSound(sampleRate, durationSeconds) {
             sin(frequency * 2 * Math.PI * it / sampleRate)
         }
